Sort by monitored project and start dashbaord

+++ start.js
@@ -19,10 +19,10 @@ async function start() {
       owner: upstream.org,
       repo: upstream.repo,
     });
-    console.log('==== ISSUES ====');
-    issues.forEach((issue) => {
-      console.log(issue.title);
-    });
+    // console.log('==== ISSUES ====');
+    // issues.forEach((issue) => {
+    //   console.log(issue.title);
+    // });
     //
     console.log(`getting commits from GitHub for: ${upstream.org}/${upstream.repo} on branch ${upstream.branch}`);
     const commits = await octokit.paginate('GET /repos/{owner}/{repo}/commits', {
@@ -30,11 +30,11 @@ async function start() {
       repo: upstream.repo,
       sha: upstream.branch,
     });
-    console.log('==== COMMITS ====');
-    commits.forEach((commit) => {
+    // console.log('==== COMMITS ====');
+    // commits.forEach((commit) => {
       // console.log(commit);
-      console.log(commit.commit.message);
-    });
+      // console.log(commit.commit.message);
+    // });
     // TODO: compare commits and issues to last
     // TODO: write last commits and issues to markdown files
 
@@ -57,31 +57,29 @@ async function start() {
     // write each commit to a .md file in content/commits
     // write each issue to a .md file in content/issues
     commits.forEach((commit) => {
-      console.log(commit);
+      // console.log(commit);
       const commitFile = `${commitsDir}/${commit.sha}.md`;
-      fs.writeFileSync(
-        commitFile,
-        `
-        ---\n
-        date: "${commit.commit.committer.date}"\n
-        reviewed: false\n
-        ---\n
-        ${commit.commit.message}
-        `,
-      );
+      if (!fs.existsSync(commitFile)) {
+        fs.writeFileSync(
+          commitFile,
+          `---\norg: "${upstream.org}"\nrepo: "${upstream.repo}"\ndate: "${commit.commit.committer.date}"\nreviewed: false\ninprogress: false\n---\n${commit.commit.message}`,
+        );
+        console.log(`wrote commit file: ${commitFile}`);
+      } else {
+        console.log(`commit file already exists: ${commitFile} -- SKIPPING`);
+      }
     });
     issues.forEach((issue) => {
       const issueFile = `${issuesDir}/${issue.number}.md`;
-      fs.writeFileSync(
-        issueFile,
-        `
-        ---\n
-        title: "${issue.title}"\n
-        reviewed: false\n
-        ---\n
-        ${issue.body}
-        `,
-      );
+      if (!fs.existsSync(issueFile)) {
+        fs.writeFileSync(
+          issueFile,
+          `---\norg: "${upstream.org}"\nrepo: "${upstream.repo}"\ntitle: "${issue.title}"\nreviewed: false\ninprogress: false\n---${issue.body}`,
+        );
+        console.log(`wrote issue file: ${issueFile}`);
+      } else {
+        console.log(`issue file already exists: ${issueFile} -- SKIPPING`);
+      }
     });
   }
 }
